[WIP]
Introduce TrickType + HandType
TrickType contains length of trick and basic trick, but not the exact cards used for trick
HandType has those as extra information.
Start implementing hand_type() function on Hand.
combinations < 5 are done, as well as FullHouse detection. Code for street exists but needs to be implemented in main function.

Street detection is done with a 4KB table