Update to a more efficient binary format for round log
Cuts down size by factor of 3 approx.
=> 7GB fit now nicely into RAM.
Ofc we could do it even more space efficient but this is a good compromise between speed and space.
Now we first unpack into a Trick object and iterate from there.

For extra speed we could go without unpacking on the binary format but the current way is conceptually much easier.